Reuse existing allOf tests and expected outputs
