--- tests/data/expected/main/jsonschema/allof_outer_constraints/description_False_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  description.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel, constr
-
-
-class Root(RootModel[constr(min_length=1)]):
-    root: constr(min_length=1) = Field('base', description='Outer description')
-
-
-class Base(RootModel[str]):
-    root: str = 'base'
