--- tests/data/expected/main/jsonschema/allof_outer_constraints/length_False_none.py
@@ -1,16 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  length.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel, constr
-
-
-class Root(RootModel[constr(min_length=1, max_length=3)]):
-    root: constr(min_length=1, max_length=3) = Field(
-        'base', description='Outer description'
-    )
-
-
-class Base(RootModel[str]):
-    root: str = 'base'
