--- tests/data/expected/main/jsonschema/allof_outer_constraints/length_True_none.py
@@ -1,16 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  length.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[str]):
-    root: str = Field(
-        'base', description='Outer description', max_length=3, min_length=1
-    )
-
-
-class Base(RootModel[str]):
-    root: str = 'base'
