--- tests/data/expected/main/jsonschema/allof_outer_constraints/minimum_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  minimum.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[str]):
-    root: str = Field(..., min_length=3)
-
-
-class Base(RootModel[str]):
-    root: str
