--- tests/data/expected/main/jsonschema/allof_outer_constraints/multiple_False_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  multiple.json
-
-from __future__ import annotations
-
-from pydantic import RootModel, conint
-
-
-class Root(RootModel[conint(multiple_of=6)]):
-    root: conint(multiple_of=6)
-
-
-class Base(RootModel[int]):
-    root: int
