--- tests/data/expected/main/jsonschema/allof_outer_constraints/multiple_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  multiple.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[int]):
-    root: int = Field(..., multiple_of=6)
-
-
-class Base(RootModel[int]):
-    root: int
