--- tests/data/expected/main/jsonschema/allof_outer_constraints/number_False_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  number.json
-
-from __future__ import annotations
-
-from pydantic import RootModel, conint
-
-
-class Root(RootModel[conint(ge=0, le=5)]):
-    root: conint(ge=0, le=5)
-
-
-class Base(RootModel[conint(le=20)]):
-    root: conint(le=20)
