--- tests/data/expected/main/jsonschema/allof_outer_constraints/number_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  number.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[int]):
-    root: int = Field(..., ge=0, le=5)
-
-
-class Base(RootModel[int]):
-    root: int = Field(..., le=20)
