--- tests/data/expected/main/jsonschema/allof_outer_constraints/ordinary_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  ordinary.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[str]):
-    root: str = Field(..., min_length=1)
-
-
-class Base(RootModel[str]):
-    root: str
