--- tests/data/expected/main/jsonschema/allof_outer_constraints/redundant_False_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  redundant.json
-
-from __future__ import annotations
-
-from pydantic import RootModel, constr
-
-
-class Root(RootModel[constr(min_length=1, max_length=3)]):
-    root: constr(min_length=1, max_length=3)
-
-
-class Base(RootModel[str]):
-    root: str
