--- tests/data/expected/main/jsonschema/allof_outer_constraints/redundant_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  redundant.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[str]):
-    root: str = Field(..., max_length=3, min_length=1)
-
-
-class Base(RootModel[str]):
-    root: str
