--- tests/data/expected/main/jsonschema/allof_outer_constraints/same_pattern_False_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  same_pattern.json
-
-from __future__ import annotations
-
-from pydantic import RootModel, constr
-
-
-class Root(RootModel[constr(pattern=r'a', min_length=1)]):
-    root: constr(pattern=r'a', min_length=1)
-
-
-class Base(RootModel[constr(pattern=r'a')]):
-    root: constr(pattern=r'a')
