--- tests/data/expected/main/jsonschema/allof_outer_constraints/same_pattern_True_none.py
@@ -1,14 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  same_pattern.json
-
-from __future__ import annotations
-
-from pydantic import Field, RootModel
-
-
-class Root(RootModel[str]):
-    root: str = Field(..., min_length=1, pattern='a')
-
-
-class Base(RootModel[str]):
-    root: str = Field(..., pattern='a')
